chore: create separate pages for interview questions and prompts
